+++ src/diffbot_system.cpp
@@ -43,7 +43,7 @@ void HardwarePub::publishData(double v, double t)
   message.temp = t;
   status_pub_->publish(message);
 }
-void HardwarePub::publishDebugData(int64_t sl, int64_t sr, int64_t cl, int64_t cr, int64_t clx, int64_t crx)
+void HardwarePub::publishDebugData(int16_t sl, int16_t sr, int16_t cl, int16_t cr, int16_t clx, int16_t crx)
 {
   auto message = robot_interfaces::msg::MotorDebug();
   message.speedr = sr;
